requirements.txt
streamlit
youtube-transcript-api
openai
faiss-cpu
numpy